JBIDE-22325 - Move job to disable all widgets into foundation, and allow astools to use it
Further changes after api refactor
