+++ plugins/org.jboss.tools.openshift.express.ui/META-INF/MANIFEST.MF
@@ -52,7 +52,8 @@ Require-Bundle: org.jboss.tools.openshift.express.client;bundle-version="[3.0.0,
  org.eclipse.jsch.ui;bundle-version="1.1.300",
  org.eclipse.jdt.debug.ui;bundle-version="3.6.1",
  org.eclipse.team.core;bundle-version="3.7.0",
- org.eclipse.swt;bundle-version="3.104.1"
+ org.eclipse.swt;bundle-version="3.104.1",
+ org.jboss.tools.foundation.ui
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.apache.commons.io;version="2.0.1",
